refactor(CalibrationUpdate): changed constants to avoid circular dependance
