+++ antropometria/config/constants/general.py
@@ -1,0 +1,32 @@
+import os
+
+from sklearn.model_selection import StratifiedKFold
+
+N_SPLITS = 10
+
+CV = StratifiedKFold(n_splits=N_SPLITS)
+FIELDNAMES = [
+    'biblioteca',
+    'classifier',
+    'reduction',
+    'filtro',
+    'min_max',
+    'balanceamento',
+    'cv_accuracy',
+    'cv_precision',
+    'cv_recall',
+    'cv_f1score',
+    'err_accuracy',
+    'err_precision_micro',
+    'err_recall_micro',
+    'err_f1score_micro',
+    'err_f1micro_ic',
+    'err_precision_macro',
+    'err_recall_macro',
+    'err_f1score_macro',
+    'err_f1macro_ic',
+    'parameters',
+]
+ROOT_DIR = os.path.abspath(os.getcwd())
+TEMPORARY_RANDOM_SAMPLES = 'antropometria/output/temp_removed_random_samples.csv'
+TEMPORARY_RANDOM_SAMPLES_LABELS = 'antropometria/output/temp_removed_random_samples_labels.csv'
